Retrosheet game logs

Retrosheet game logs for all games played in the 2011 season

Usage

retro_gl_header

retro_gl_2011

Format

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame) with 0 rows and 161 columns.

Date

date in the form yyyymmdd

DoubleHeader

number of game

DayOfWeek

day of week

VisitingTeam

visiting team

VisitingTeamLeague

visiting league

VisitingTeamGameNumber

visiting team game number

HomeTeam

home team

HomeTeamLeague

home league

HomeTeamGameNumber

home team game number

VisitorRunsScored

visiting score

HomeRunsScore

home score

LengthInOuts

length of game in outs

DayNight

day/night indicator

CompletionInfo

completed game information

ForfeitInfo

forfeit game information

ProtestInfo

protest information

ParkID

park id

Attendence

attendance

Duration

time of game in minutes

VisitorLineScore

visiting team line score

HomeLineScore

home team line score

VisitorAB

visiting team at-bats

VisitorH

visiting team hits

VisitorD

visiting team doubles

VisitorT

visiting team triples

VisitorHR

visiting team home runs

VisitorRBI

visiting team runs batted in

VisitorSH

visiting team sacrifice hits

VisitorSF

visiting team sacrifice flies

VisitorHBP

visiting team hit by pitches

VisitorBB

visiting team walks

VisitorIBB

visiting team intentional walks

VisitorK

visiting team strikeouts

VisitorSB

visiting team stolen bases

VisitorCS

visiting team caught stealing

VisitorGDP

visiting team ground in double plays

VisitorCI

visiting team awarded first on catcher interference

VisitorLOB

visiting team left on base

VisitorPitchers

visiting team pitchers used

VisitorER

visiting team individual earned runs

VisitorTER

visiting team team earned runs

VisitorWP

visiting team wild pitches

VisitorBalks

visiting team balks

VisitorPO

visiting team putouts

VisitorA

visiting team assists

VisitorE

visiting team errors

VisitorPassed

visiting team passed balls

VisitorDB

visiting team double plays

VisitorTP

visiting team triple plays

HomeAB

home team at bats

HomeH

home team hits

HomeD

home team doubles

HomeT

home team triples

HomeHR

home team home runs

HomeRBI

home team runs batted in

HomeSH

home team sacrifice hits

HomeSF

home team sacrifice flies

HomeHBP

home team hit by pitches

HomeBB

home team walks

HomeIBB

home team intentional walks

HomeK

home team strikeouts

HomeSB

home team stolen bases

HomeCS

home team caught stealing

HomeGDP

home team ground in double plays

HomeCI

home team awarded first on catcher interference

HomeLOB

home team left on base

HomePitchers

home team number of pitchers

HomeER

home team individual earned runs

HomeTER

home team team earned runs

HomeWP

home team wild pitches

HomeBalks

home team balks

HomePO

home team putouts

HomeA

home team assists

HomeE

home team errors

HomePassed

home team passed balls

HomeDB

home team double plays

HomeTP

home team triple plays

UmpireHID

home plate umpire id

UmpireHName

home plate umpire name

Umpire1BID

1st base umpire id

Umpire1BName

1st base umpire name

Umpire2BID

2nd base umpire id

Umpire2BName

2nd base umpire name

Umpire3BID

3rd base umpire id

Umpire3BName

3rd base umpire name

UmpireLFID

LF umpire id

UmpireLFName

LF umpire name

UmpireRFID

RF umpire id

UmpireRFName

RF umpire name

VisitorManagerID

visiting manager id

VisitorManagerName

visiting manager name

HomeManagerID

home manager id

HomeManagerName

home manager name

WinningPitcherID

winning pitcher id

WinningPitcherName

winning pitcher name

LosingPitcherID

losing pitcher id

LosingPitcherNAme

losing pitcher name

SavingPitcherID

saving pitcher id

SavingPitcherName

saving pitcher name

GameWinningRBIID

id of player with the winning RBI

GameWinningRBIName

name of player with the winning RBI

VisitorStartingPitcherID

visiting starting pitcher id

VisitorStartingPitcherName

visiting starting pitcher name

HomeStartingPitcherID

home starting pitcher id

HomeStartingPitcherName

home starting pitcher name

VisitorBatting1PlayerID

visiting batting 1st id

VisitorBatting1Name

visiting batting 1st name

VisitorBatting1Position

visiting batting 1st position

VisitorBatting2PlayerID

visiting batting 2nd name

VisitorBatting2Name

visiting batting 2nd id

VisitorBatting2Position

visiting batting 2nd position

VisitorBatting3PlayerID

visiting batting 3rd id

VisitorBatting3Name

visiting batting 3rd name

VisitorBatting3Position

visiting batting 3rd position

VisitorBatting4PlayerID

visiting batting 4th id

VisitorBatting4Name

visiting batting 4th name

VisitorBatting4Position

visiting batting 4th position

VisitorBatting5PlayerID

visiting batting 5th id

VisitorBatting5Name

visiting batting 5th name

VisitorBatting5Position

visiting batting 5th position

VisitorBatting6PlayerID

visiting batting 6th id

VisitorBatting6Name

visiting batting 6th name

VisitorBatting6Position

visiting batting 6th position

VisitorBatting7PlayerID

visiting batting 7th id

VisitorBatting7Name

visiting batting 7th name

VisitorBatting7Position

visiting batting 7th position

VisitorBatting8PlayerID

visiting batting 8th id

VisitorBatting8Name

visiting batting 8th name

VisitorBatting8Position

visiting batting 8th position

VisitorBatting9PlayerID

visiting batting 9th id

VisitorBatting9Name

visiting batting 9th name

VisitorBatting9Position

visiting batting 9th position

HomeBatting1PlayerID

home batting 1st id

HomeBatting1Name

home batting 1st name

HomeBatting1Position

home batting 1st position

HomeBatting2PlayerID

home batting 2nd id

HomeBatting2Name

home batting 2nd name

HomeBatting2Position

home batting 2nd position

HomeBatting3PlayerID

home batting 3rd id

HomeBatting3Name

home batting 3rd name

HomeBatting3Position

home batting 3rd position

HomeBatting4PlayerID

home batting 4th id

HomeBatting4Name

home batting 4th name

HomeBatting4Position

home batting 4th position

HomeBatting5PlayerID

home batting 5th id

HomeBatting5Name

home batting 5th name

HomeBatting5Position

home batting 5th position

HomeBatting6PlayerID

home batting 6th id

HomeBatting6Name

home batting 6th name

HomeBatting6Position

home batting 6th position

HomeBatting7PlayerID

home batting 7th id

HomeBatting7Name

home batting 7th name

HomeBatting7Position

home batting 7th position

HomeBatting8PlayerID

home batting 8th id

HomeBatting8Name

home batting 8th name

HomeBatting8Position

home batting 8th position

HomeBatting9PlayerID

home batting 9th id

HomeBatting9Name

home batting 9th name

HomeBatting9Position

home batting 9th position

AdditionalInfo

additional information

AcquisitionInfo

acquisition information

Details

For retro_gl_header, the data frame contains no rows, only the variable names.